requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
pydantic-settings==2.5.2
uvicorn[standard]
supabase
python-jose[cryptography]
PyJWT>=2.8.0
langchain==0.3.13
langchain-openai==0.2.14
langchain-anthropic==0.3.4
langchain-google-genai==2.0.8
langgraph==0.2.59
python-dotenv==1.0.0

# Web Crawling
crawlee[beautifulsoup]>=0.3.0
crawlee[playwright]>=0.3.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Web Search
duckduckgo-search>=4.0.0

# HTTP Client (async)
httpx>=0.25.0

# Vector Stores
pinecone-client>=5.0.0
pgvector>=0.3.6

# Academic Sources
arxiv>=2.1.0

# YouTube
youtube-transcript-api>=0.6.2
google-api-python-client>=2.147.0

# PDF Parsing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0

# Image Processing
Pillow>=10.0.0

# Chart Generation
matplotlib>=3.8.0
plotly>=5.18.0
kaleido>=0.2.1

# Translation (using LLM + MyMemory API — googletrans conflicts with httpx)

# Math & Symbolic Computation
sympy>=1.12

# Data Analysis
pandas>=2.0.0
openpyxl>=3.1.0

# Knowledge Graph (in-memory MVP)
networkx>=3.0

# Code Execution (Docker sandbox)
docker>=7.0.0